Configuration.hs: use Data.Map.{Strict,Lazy} for ghc-8.6.3
Thanks to Sergey Alirzaev for the patch!

I have tested that this continues to allow hackport to build
successfully on ghc-8.4.4, and Sergey suggests that this allows
hackport to build on ghc-8.6.3.

I tested the resulting executable by using the merge command on
pandoc and ansi-terminal. Both seemed to generate correct ebuilds.

Reported-By: Sergey Alirzaev (l29ah)
Bug: https://github.com/gentoo-haskell/hackport/issues/49
Signed-off-by: Jack Todaro <jackmtodaro@gmail.com>